[UPD] split logic from layout to a reporter
