+++ src/components/form/Form.module.scss
@@ -17,7 +17,7 @@
 }
 
 .FormMax {
-  max-width: 97vw!important;
+  max-width: 97vw !important;
 }
 
 .FormHeader {
@@ -33,12 +33,14 @@
   position: absolute;
   top: 0;
   right: 0;
+  height: 20px;
+  width: 20px;
 }
 
 .FormIcons {
   height: 20px;
   width: 20px;
-  cursor:pointer;
+  cursor: pointer;
   filter: brightness(0) invert(1);
   transition: all 0.2s ease-out;
   position: absolute;
